+++ src/main/scala/com/wenjunhuang/codeepiphany/codeforces/settings/CodeForcesSettingsConfigurable.scala
@@ -3,12 +3,12 @@ package com.wenjunhuang.codeepiphany.codeforces.settings
 import com.intellij.openapi.project.Project
 
 import com.wenjunhuang.codeepiphany.PluginBundle
-import com.wenjunhuang.codeepiphany.codeforces.models.{CodeForcesChallengeCodeTemplate, CodeForcesChallengeData}
+import com.wenjunhuang.codeepiphany.codeforces.models.{ CodeForcesChallengeCodeTemplate, CodeForcesChallengeData }
 import com.wenjunhuang.codeepiphany.codeforces.settings.CodeForcesSettingsConfigurable.*
 import com.wenjunhuang.codeepiphany.model.*
 import com.wenjunhuang.codeepiphany.model.Language.*
 import com.wenjunhuang.codeepiphany.model.LanguageVersion.*
-import com.wenjunhuang.codeepiphany.settings.dojo.{BaseCodeDojoSettings, BaseSettingsConfigurable}
+import com.wenjunhuang.codeepiphany.settings.dojo.{ BaseCodeDojoSettings, BaseSettingsConfigurable }
 
 class CodeForcesSettingsConfigurable(project: Project)
     extends BaseSettingsConfigurable(
@@ -34,33 +34,35 @@ class CodeForcesSettingsConfigurable(project: Project)
 
 object CodeForcesSettingsConfigurable {
   val CODEFORCES_LANGUAGES: Map[(Language, LanguageVersion), String] = Map(
-    (Java, SpecificVersion("8"))        -> "36",
-    (Java, SpecificVersion("21"))       -> "87",
-    (C, SpecificVersion("11"))          -> "43",
-    (Cpp, SpecificVersion("17"))        -> "54",
-    (Cpp, SpecificVersion("20"))        -> "89",
-    (Cpp, SpecificVersion("23"))        -> "91",
-    (CSharp, SpecificVersion("8"))      -> "65",
-    (CSharp, SpecificVersion("10"))     -> "79",
-    (CSharp, SpecificVersion("Mono"))   -> "9",
-    (Kotlin, SpecificVersion("1.7.20")) -> "83",
-    (Kotlin, SpecificVersion("1.9.21")) -> "88",
-    (OCaml, AnyVersion)                 -> "19",
-    (Delphi, AnyVersion)                -> "3",
-    (Pascal, AnyVersion)                -> "4",
-    (D, AnyVersion)                     -> "28",
-    (GO, AnyVersion)                    -> "32",
-    (Haskell, AnyVersion)               -> "12",
-    (Perl, AnyVersion)                  -> "13",
-    (PHP, AnyVersion)                   -> "6",
-    (Python, SpecificVersion("2"))      -> "7",
-    (Python, SpecificVersion("3"))      -> "31",
-    (Pypy, SpecificVersion("2"))        -> "40",
-    (Pypy, SpecificVersion("3"))        -> "70",
-    (Ruby, AnyVersion)                  -> "67",
-    (Rust, AnyVersion)                  -> "75",
-    (Scala, AnyVersion)                 -> "20",
-    (Javascript, AnyVersion)            -> "55"
+    (Java, SpecificVersion("8(32bit)"))                     -> "36",
+    (Java, SpecificVersion("21(64bit)"))                    -> "87",
+    (C, SpecificVersion("11(GCC 5.1.0)"))                   -> "43",
+    (Cpp, SpecificVersion("17(GCC 7.3.0)"))                 -> "54",
+    (Cpp, SpecificVersion("20(GCC 13.2)"))                  -> "89",
+    (Cpp, SpecificVersion("23(GCC 14.2)"))                  -> "91",
+    (CSharp, SpecificVersion("8(.NET Core 3.1)"))           -> "65",
+    (CSharp, SpecificVersion("10(.NET SDK 6.0)"))           -> "79",
+    (CSharp, SpecificVersion("(Mono 6.8)"))                 -> "9",
+    (Kotlin, SpecificVersion("1.7.20"))                     -> "83",
+    (Kotlin, SpecificVersion("1.9.21"))                     -> "88",
+    (OCaml, SpecificVersion("4.02.1"))                      -> "19",
+    (Delphi, SpecificVersion("7"))                          -> "3",
+    (Pascal, SpecificVersion("3.2.2(Free)"))                -> "4",
+    (D, SpecificVersion("(DMD32 v2.105.0)"))                -> "28",
+    (GO, SpecificVersion("1.22.2"))                         -> "32",
+    (Haskell, SpecificVersion("(GHC 8.10.1)"))              -> "12",
+    (Perl, SpecificVersion("5.20.1"))                       -> "13",
+    (PHP, SpecificVersion("8.1.7"))                         -> "6",
+    (Python, SpecificVersion("2.7.18"))                     -> "7",
+    (Python, SpecificVersion("3.8.10"))                     -> "31",
+    (Pypy, SpecificVersion("2.7.13(7.3.0)"))                -> "40",
+    (Pypy, SpecificVersion("3.6.9(7.3.0)"))                 -> "41",
+    (Pypy, SpecificVersion("3.10(7.3.15 64bit)"))           -> "70",
+    (Ruby, SpecificVersion("3.2.2"))                        -> "67",
+    (Rust, SpecificVersion("1.75.0(2021)"))                 -> "75",
+    (Scala, SpecificVersion("2.12.8"))                      -> "20",
+    (Javascript, SpecificVersion("(V8 4.8.0)"))             -> "34",
+    (Javascript, SpecificVersion("(Node.js 15.8.0 64bit)")) -> "55"
   )
   val CODEFORCES_LANGUAGES_REVERSE: Map[String, (Language, LanguageVersion)] = CODEFORCES_LANGUAGES.map(_.swap)