+++ src/main/scala/com/wenjunhuang/codeepiphany/toolwindows/sidebar/submissionlog/leetcode/LeetCodeSubmissionResultForm.java
@@ -5,15 +5,19 @@ import com.intellij.ide.ui.laf.darcula.ui.DarculaEditorTextFieldBorder;
 import com.intellij.openapi.editor.EditorFactory;
 import com.intellij.openapi.editor.colors.EditorFontType;
 import com.intellij.openapi.fileTypes.FileTypeManager;
+import com.intellij.openapi.util.text.StringUtil;
+import com.intellij.ui.BrowserHyperlinkListener;
 import com.intellij.ui.EditorTextField;
 import com.intellij.uiDesigner.core.GridConstraints;
 import com.intellij.uiDesigner.core.GridLayoutManager;
 import com.intellij.uiDesigner.core.Spacer;
+import com.intellij.util.ui.HTMLEditorKitBuilder;
 import com.intellij.util.ui.JBFont;
 import com.intellij.util.ui.JBUI;
 import com.wenjunhuang.codeepiphany.PluginBundle;
 import com.wenjunhuang.codeepiphany.database.tables.records.LeetcodeSubmissionRecord;
 import com.wenjunhuang.codeepiphany.database.tables.records.SolutionSubmissionRecord;
+import com.wenjunhuang.codeepiphany.model.CodeDojo;
 import com.wenjunhuang.codeepiphany.model.Language;
 import com.wenjunhuang.codeepiphany.model.LanguageVersion;
 import com.wenjunhuang.codeepiphany.model.SubmissionResult;
@@ -35,15 +39,17 @@ public class LeetCodeSubmissionResultForm {
     private JLabel myTestcases;
     private JLabel mySubmitDateTime;
     private JLabel myLanguage;
-    private JLabel myLanguageVersion;
     private EditorTextField mySubmitCode;
     private JPanel myPanel;
     private JComponent myResultComponent;
     private JLabel myCodeLabel;
+    private JEditorPane myViewInBrowser;
 
     public LeetCodeSubmissionResultForm(
             Language language,
             LanguageVersion languageVersion,
+            String challengeSlug,
+            CodeDojo leetCodeDojo,
             SolutionSubmissionRecord submissionRecord,
             LeetcodeSubmissionRecord leetcodeSubmissionRecord
     ) {
@@ -86,6 +92,17 @@ public class LeetCodeSubmissionResultForm {
                         JBUI.Borders.customLine(JBUI.CurrentTheme.CustomFrameDecorations.separatorForeground(), 0, 1, 0, 0),
                         JBUI.Borders.emptyLeft(2)
                 ));
+        var submissionId = submissionRecord.getDojosubmissionid();
+        if (submissionId != null) {
+            var link = "https://" + leetCodeDojo.domain().toString() + "/problems/" + challengeSlug + "/submissions/" + submissionId + "/";
+
+            myViewInBrowser.setEditorKit(HTMLEditorKitBuilder.simple());
+            myViewInBrowser.setEditable(false);
+            myViewInBrowser.addHyperlinkListener(new BrowserHyperlinkListener());
+            myViewInBrowser.setText("<html><body><a href='" + link + "'>View In Browser</a></body></html>");
+        } else {
+            myViewInBrowser.setVisible(false);
+        }
     }
 
     public JComponent getComponent() {
@@ -105,7 +122,7 @@ public class LeetCodeSubmissionResultForm {
     private void $$$setupUI$$$() {
         createUIComponents();
         myPanel = new JPanel();
-        myPanel.setLayout(new GridLayoutManager(5, 3, new Insets(0, 0, 0, 0), 2, 5));
+        myPanel.setLayout(new GridLayoutManager(5, 4, new Insets(0, 0, 0, 0), 2, 5));
         myPanel.setBorder(BorderFactory.createTitledBorder(BorderFactory.createEmptyBorder(0, 5, 0, 5), null, TitledBorder.DEFAULT_JUSTIFICATION, TitledBorder.DEFAULT_POSITION, null, null));
         myTestcases = new JLabel();
         myTestcases.setText("Testcase");
@@ -114,16 +131,16 @@ public class LeetCodeSubmissionResultForm {
         myResult.setText("Label");
         myPanel.add(myResult, new GridConstraints(0, 0, 1, 1, GridConstraints.ANCHOR_WEST, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_FIXED, GridConstraints.SIZEPOLICY_FIXED, null, null, null, 0, false));
         final Spacer spacer1 = new Spacer();
-        myPanel.add(spacer1, new GridConstraints(0, 2, 1, 1, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_HORIZONTAL, GridConstraints.SIZEPOLICY_WANT_GROW, 1, null, null, null, 0, false));
+        myPanel.add(spacer1, new GridConstraints(0, 3, 1, 1, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_HORIZONTAL, GridConstraints.SIZEPOLICY_WANT_GROW, 1, null, null, null, 0, false));
         mySubmitDateTime = new JLabel();
         mySubmitDateTime.setText("Label");
-        myPanel.add(mySubmitDateTime, new GridConstraints(1, 0, 1, 3, GridConstraints.ANCHOR_WEST, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_FIXED, GridConstraints.SIZEPOLICY_FIXED, null, null, null, 0, false));
+        myPanel.add(mySubmitDateTime, new GridConstraints(1, 0, 1, 4, GridConstraints.ANCHOR_WEST, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_FIXED, GridConstraints.SIZEPOLICY_FIXED, null, null, null, 0, false));
         mySubmitCode.setViewer(false);
-        myPanel.add(mySubmitCode, new GridConstraints(4, 0, 1, 3, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_BOTH, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_WANT_GROW, null, null, null, 0, false));
-        myPanel.add(myResultComponent, new GridConstraints(2, 0, 1, 3, GridConstraints.ANCHOR_NORTH, GridConstraints.FILL_HORIZONTAL, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_FIXED, null, null, null, 0, false));
+        myPanel.add(mySubmitCode, new GridConstraints(4, 0, 1, 4, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_BOTH, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_WANT_GROW, null, null, null, 0, false));
+        myPanel.add(myResultComponent, new GridConstraints(2, 0, 1, 4, GridConstraints.ANCHOR_NORTH, GridConstraints.FILL_HORIZONTAL, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_FIXED, null, null, null, 0, false));
         final JPanel panel1 = new JPanel();
         panel1.setLayout(new GridLayoutManager(1, 3, new Insets(0, 0, 0, 0), -1, -1));
-        myPanel.add(panel1, new GridConstraints(3, 0, 1, 3, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_BOTH, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, null, null, null, 0, false));
+        myPanel.add(panel1, new GridConstraints(3, 0, 1, 4, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_BOTH, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, null, null, null, 0, false));
         myCodeLabel = new JLabel();
         this.$$$loadLabelText$$$(myCodeLabel, this.$$$getMessageFromBundle$$$("messages/PluginBundle", "leetcode.submissionResult.code.text"));
         panel1.add(myCodeLabel, new GridConstraints(0, 0, 1, 1, GridConstraints.ANCHOR_WEST, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_FIXED, GridConstraints.SIZEPOLICY_FIXED, null, null, null, 0, false));
@@ -132,6 +149,8 @@ public class LeetCodeSubmissionResultForm {
         myLanguage = new JLabel();
         myLanguage.setText("Label");
         panel1.add(myLanguage, new GridConstraints(0, 1, 1, 1, GridConstraints.ANCHOR_WEST, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_FIXED, GridConstraints.SIZEPOLICY_FIXED, null, null, null, 0, false));
+        myViewInBrowser = new JEditorPane();
+        myPanel.add(myViewInBrowser, new GridConstraints(0, 2, 1, 1, GridConstraints.ANCHOR_WEST, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_WANT_GROW, GridConstraints.SIZEPOLICY_WANT_GROW, null, null, null, 0, false));
     }
 
     private static Method $$$cachedGetBundleMethod$$$ = null;